Move OriginalMapFileDataStructs subclasses to their own files
